Move some tests to the core module and do some manual testing in testapp
