+++ app/build.gradle
@@ -53,28 +53,6 @@ dependencies {
     androidTestImplementation('junit:junit:4.13.2')
     androidTestImplementation 'androidx.test:rules:1.5.0'
     androidTestImplementation 'org.easytesting:fest-util:1.2.5'
-
-    testImplementation 'org.hamcrest:hamcrest-core:1.3'
-    testImplementation 'org.hamcrest:hamcrest-library:1.3'
-    testImplementation('junit:junit:4.13.2')
-    testImplementation 'org.easytesting:fest-util:1.2.5'
-    testImplementation('org.robolectric:robolectric:4.11.1') {
-        exclude module: 'classworlds'
-        exclude module: 'commons-logging'
-        exclude module: 'httpclient'
-        exclude module: 'maven-artifact'
-        exclude module: 'maven-artifact-manager'
-        exclude module: 'maven-error-diagnostics'
-        exclude module: 'maven-model'
-        exclude module: 'maven-project'
-        exclude module: 'maven-settings'
-        exclude module: 'plexus-container-default'
-        exclude module: 'plexus-interpolation'
-        exclude module: 'plexus-utils'
-        exclude module: 'wagon-file'
-        exclude module: 'wagon-http-lightweight'
-        exclude module: 'wagon-provider-api'
-    }
 }
 
 publishing {
