+++ core/build.gradle
@@ -42,7 +42,9 @@ android {
 dependencies {
     api 'org.slf4j:slf4j-api:1.7.36'
 
-    testImplementation 'junit:junit:4.13.2'
+    testImplementation('junit:junit:4.13.2')
+    testImplementation 'org.hamcrest:hamcrest-core:1.3'
+    testImplementation 'org.hamcrest:hamcrest-library:1.3'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
