+++ core/src/test/java/pl/brightinventions/slf4android/HandlerFormatterCompilerTests.java
